requirements.txt
Django==5.0.6
djangorestframework==3.15.1
redis==5.0.4
python-dotenv==1.0.1

# Dev / test
pytest==8.2.0
pytest-django==4.8.0
python-decouple==3.8